Fixed depth testing

+++ include/datagui/geometry_renderer.hpp
@@ -12,8 +12,8 @@ public:
     GeometryRenderer();
     void init();
 
-    void queue_box(int depth, const Boxf& box, const Color& color, float radius = 0.f);
-    void queue_box(int depth, const Boxf& box, const Color& color, float border_radius, Color border_color, float radius = 0.f);
+    void queue_box(float depth, const Boxf& box, const Color& color, float radius = 0.f);
+    void queue_box(float depth, const Boxf& box, const Color& color, float border_radius, Color border_color, float radius = 0.f);
 
     void render(const Vecf& viewport_size);
